Added files via upload
IMPORTANT this method takes in an int[] indicating the position being placed and a String whoAmI for who is placing.
Signature as follows ---	public void flipThatShit(int[] placeHere, String whoAmI){

Okay, so I added the flip tiles method. It shouldn't be too difficult to reimplement the method to give the hypothetical number of tiles flipped given a point to do the intelligentComputer thingy. I haven't tested this at all and am not really sure how to, but it probably works :/ so yeah.